app-compat (Track B/B7): end-to-end multi-DSO loading + PAGE_MASK bugfix
tobyOS now runs Alpine's real `file` binary, which depends on a SEPARATE
shared library (libmagic.so.1, beyond libc). `file --version` makes the real
musl ld.so open /lib/libmagic.so.1 and file-backed-mmap its segments (one
whole-span MAP_PRIVATE reservation, then a MAP_FIXED file map per segment at
the library's base + offset), relocate it, resolve its symbols, and print
"file-5.45": [LXMULTI] VERDICT: PASS, clean under default boot AND
-cpu qemu64,+smep,+smap (0 faults, 0 unhandled), 3x100s validate ALIVE.

Root-cause bugfix (the real blocker behind the earlier multi-DSO faults):
src/mmap.c defined PAGE_MASK as ~(PAGE_SIZE - 1), and PAGE_SIZE comes from
pmm.h as 4096u (32-bit unsigned), so the mask was 0x00000000FFFFF000 --
page_align_down/page_align_up silently TRUNCATED any address >= 4 GiB. The
mmap region lives at 16 TiB (MMAP_REGION_BASE), so ld-musl's MAP_FIXED
per-segment maps (e.g. 0x100000018000) were aligned down to 0x18000, and the
loader then wrote relocations to the wrong page and NULL-deref'd. Fixed to
~((uint64_t)PAGE_SIZE - 1). This was latent: anonymous mmap uses
find_free_region (not page_align_down), and only MAP_FIXED / munmap /
mprotect on the 16 TiB region truncated -- paths nothing exercised before
the Linux dynamic loader. (So B6's file-backed mmap was correct all along.)

Diagnosed with a temporary -DLXMMAP_TRACE mmap tracer (kept in LX_mmap, off
by default) that printed "addr=0x100000018000 ... -> 0x18000", immediately
revealing the truncation.

The multi-DSO `file` + libmagic.so.1 are a real Alpine package, opt-in and
NOT committed (.gitignore; programs/busybox/README.md documents the fetch).
This commit is the PAGE_MASK fix + the mmap tracer + the LXMULTI boot
harness + opt-in initrd staging + docs.

Deferred to B8: fork/clone/execve/wait4 (for `busybox sh` with external
commands), threads (clone), and SA_SIGINFO 3-arg handlers.

+++ tobyOS/src/mmap.c
@@ -66,7 +66,13 @@ static struct vma_table g_vma_tables[PROC_MAX];
 #ifndef PAGE_SIZE
 #define PAGE_SIZE 4096ULL
 #endif
-#define PAGE_MASK (~(PAGE_SIZE - 1))
+/* PAGE_SIZE comes from pmm.h as `4096u` (32-bit unsigned), so the naive
+ * `~(PAGE_SIZE - 1)` is a 32-bit `0xFFFFF000` that zero-extends to
+ * `0x00000000FFFFF000` and TRUNCATES any address >= 4 GiB -- and the mmap
+ * region lives at 16 TiB (MMAP_REGION_BASE). Force the mask 64-bit so
+ * page_align_down/up keep the high bits (a real bug the Linux dynamic
+ * loader's MAP_FIXED segment maps surfaced). */
+#define PAGE_MASK (~((uint64_t)PAGE_SIZE - 1))
 
 static inline uint64_t page_align_up(uint64_t v) {
     return (v + PAGE_SIZE - 1) & PAGE_MASK;
